requirements.txt
pyrogram
rich
aiogram
aiofiles
pysocks
python-socks